requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
selenium>=4.8.0
chromedriver-autoinstaller>=0.6.0
lxml>=4.9.0
fake-useragent>=1.2.0
pysocks>=1.7.1
PyYAML>=6.0
schedule>=1.2.0
